+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KRBSBMVJPA38F2D5QSRPBDCT"
+        "01KRBW9CH1Y8A360Y05PXXNDX6"
     )
     r"""The unique identifier of the data source"""