+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KRBSBMVQRR3YHANT6R675Q4Z"
+        "01KRBW9CHBC113MRF5HRGFR7MY"
     )
     r"""The unique identifier of the data source"""